WIP: refactor tests into new file `test_io_odim.py`, minor fixes
